Tests/LibWeb/Text/expected/keypress-on-focused-button-triggers-onclick.txt

<BUTTON id="button1">
Clicked button 1
<BUTTON id="button2">
Clicked button 2
